fix: refactor files, fix switch component, fix debouncing bug
